refactor: fix typos in yaml-editor and lunar-phase-card imports, update deprecated theme config, and improve chunk-object iteration
