+++ src/utils/compute-stub-config.ts
